add cast to fix buildbot
Reviewed-on: https://code.wireshark.org/review/2476

+++ epan/dissectors/packet-corosync-totemnet.c
@@ -311,7 +311,7 @@ dissect_corosynec_totemnet_with_decryption(tvbuff_t *tvb,
   memset (keys, 0, sizeof(keys));
   sober128_start (&keygen_prng_state);
   sober128_add_entropy(private_key,
-                                  private_key_len, &keygen_prng_state);
+                                  (unsigned long)private_key_len, &keygen_prng_state);
   sober128_add_entropy (salt, SALT_SIZE, &keygen_prng_state);
   sober128_read (keys, sizeof (keys), &keygen_prng_state);
